+++ readme.md
@@ -71,4 +71,5 @@
   - [x] 회원가입
   - [x] 도서관리
   - [x] 도서 대출/반납
+  - [ ] BeanValidation 적용
 - [x] 관리자 페이지 세분화 (도서관리, 회원관리 나눔)
