+++ src/main/java/com/weblibrary/domain/user/dto/LoginUserDto.java
@@ -1,0 +1,16 @@
+package com.weblibrary.domain.user.dto;
+
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.Size;
+import lombok.Data;
+
+@Data
+public class LoginUserDto {
+    @NotBlank
+    @Size(min = 5)
+    private String username;
+
+    @NotBlank
+    @Size(min = 5)
+    private String password;
+}
