--- src/main/java/com/weblibrary/domain/user/model/JoinUserDto.java
@@ -1,13 +1,0 @@
-package com.weblibrary.domain.user.model;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class JoinUserDto {
-    private String username;
-    private String password;
-}
