--- src/main/java/com/weblibrary/domain/user/model/LoginUserDto.java
@@ -1,9 +1,0 @@
-package com.weblibrary.domain.user.model;
-
-import lombok.Data;
-
-@Data
-public class LoginUserDto {
-    private String username;
-    private String password;
-}
