+++ src/main/java/com/weblibrary/domain/user/service/UserService.java
@@ -3,8 +3,8 @@ package com.weblibrary.domain.user.service;
 import com.weblibrary.domain.admin.model.Role;
 import com.weblibrary.domain.admin.repository.MemoryUserRoleRepository;
 import com.weblibrary.domain.admin.repository.UserRoleRepository;
-import com.weblibrary.domain.user.model.JoinUserDto;
-import com.weblibrary.domain.user.model.LoginUserDto;
+import com.weblibrary.domain.user.dto.JoinUserDto;
+import com.weblibrary.domain.user.dto.LoginUserDto;
 import com.weblibrary.domain.user.model.User;
 import com.weblibrary.domain.user.repository.MemoryUserRepository;
 import com.weblibrary.domain.user.repository.UserRepository;
