+++ src/main/resources/errors.properties
@@ -1,13 +1,9 @@
 loginGlobal=\uB85C\uADF8\uC778\uC5D0 \uC2E4\uD328\uD588\uC2B5\uB2C8\uB2E4. \uC544\uC774\uB514 \uBC0F \uD328\uC2A4\uC6CC\uB4DC\uB97C \uD655\uC778\uD558\uC138\uC694.
-required.user.username=\uC720\uC800\uC774\uB984\uC740 \uD544\uC218\uC785\uB2C8\uB2E4.
-required.user.password=\uD328\uC2A4\uC6CC\uB4DC\uB294 \uD544\uC218\uC785\uB2C8\uB2E4.
 required.book.bookName=\uCC45 \uC774\uB984\uC740 \uD544\uC218\uC785\uB2C8\uB2E4.
 required.book.isbn=ISBN\uC740 \uD544\uC218\uC785\uB2C8\uB2E4.
 required=\uD544\uC218 \uAC12\uC785\uB2C8\uB2E4.
 
 duplicated.user.username=\uC911\uBCF5\uB41C \uC720\uC800 \uC774\uB984\uC785\uB2C8\uB2E4.
-min.user.username=\uC720\uC800\uC774\uB984\uC740 {0}\uC790 \uC774\uC0C1\uC73C\uB85C \uD574\uC8FC\uC138\uC694.
-min.user.password=\uD328\uC2A4\uC6CC\uB4DC\uB294 {0}\uC790 \uC774\uC0C1\uC73C\uB85C \uD574\uC8FC\uC138\uC694.
 min.book.bookName=\uCC45 \uC774\uB984\uC740 \uCD5C\uC18C {0}\uC790 \uC774\uC0C1 \uC785\uB825\uD558\uC138\uC694.
 min.book.isbn=isbn\uC740 \uCD5C\uC18C {0}\uC790 \uC774\uC0C1 \uC785\uB825\uD558\uC138\uC694.
 min=\uCD5C\uC18C {0}\uC790 \uC774\uC0C1\uC73C\uB85C \uD574\uC8FC\uC138\uC694.
@@ -22,3 +18,9 @@ invalid.status.rentBookDto.book=\uB300\uCD9C \uC0C1\uD0DC\uAC00 \uC544\uB2CC \uB
 invalid.rented.rentBookDto.book=\uBE4C\uB9AC\uC9C0 \uC54A\uC740 \uB3C4\uC11C\uC785\uB2C8\uB2E4.
 invalid.available.rentBookDto.book=\uC774\uBBF8 \uB300\uCD9C\uC911\uC778 \uB3C4\uC11C\uC785\uB2C8\uB2E4. \uB300\uCD9C\uC911\uC778 \uC720\uC800\uB294 {0}\uB2D8\uC785\uB2C8\uB2E4.
 invalid.remain.rentBookDto.user=\uB300\uCD9C \uAC00\uB2A5 \uD69F\uC218\uAC00 \uCD08\uACFC\uB418\uC5C8\uC2B5\uB2C8\uB2E4. \uB0A8\uC740 \uB300\uCD9C \uD69F\uC218\uB294 {0}\uD68C \uC785\uB2C8\uB2E4.
+
+Size.user.username=\uC720\uC800\uC774\uB984\uC740 {2}\uC790 \uC774\uC0C1\uC774\uC5B4\uC5EC \uD569\uB2C8\uB2E4.
+Size.user.password=\uD53C\uC2A4\uC6CC\uB4DC\uB294 {2}\uC790 \uC774\uC0C1\uC774\uC5B4\uC5EC \uD569\uB2C8\uB2E4.
+
+NotBlank.user.username=\uC720\uC800\uC774\uB984\uC740 \uD544\uC218\uC785\uB2C8\uB2E4.
+NotBlank.user.password=\uD328\uC2A4\uC6CC\uB4DC\uB294 \uD544\uC218\uC785\uB2C8\uB2E4.
